refactor(apidom-ls): reorganize existing bindings rules
Refs #1647
