+++ packages/apidom-ls/src/config/asyncapi/bindings/http/message-binding/completion.ts
@@ -2,7 +2,7 @@ import {
   ApidomCompletionItem,
   CompletionFormat,
   CompletionType,
-} from '../../../apidom-language-types';
+} from '../../../../../apidom-language-types';
 
 const completion: ApidomCompletionItem[] = [
   {
