+++ packages/apidom-ls/src/config/asyncapi/bindings/http/message-binding/lint/bindingVersion.ts
@@ -1,0 +1,16 @@
+import ApilintCodes from '../../../../../codes';
+import { LinterMeta } from '../../../../../../apidom-language-types';
+
+const httpMessageBindingsVersionLint: LinterMeta = {
+  code: ApilintCodes.MESSAGE_BINDING_HTTP_BINDINGVERSION,
+  source: 'apilint',
+  message: "'bindingVersion' value must be a string",
+  severity: 1,
+  linterFunction: 'apilintType',
+  linterParams: ['string'],
+  marker: 'value',
+  target: 'bindingVersion',
+  data: {},
+};
+
+export default httpMessageBindingsVersionLint;
