+++ packages/apidom-ls/src/config/asyncapi/bindings/http/message-binding/lint/headers.ts
@@ -1,0 +1,16 @@
+import ApilintCodes from '../../../../../codes';
+import { LinterMeta } from '../../../../../../apidom-language-types';
+
+const httpMessageBindingHeadersLint: LinterMeta = {
+  code: ApilintCodes.MESSAGE_BINDING_HTTP_HEADERS,
+  source: 'apilint',
+  message: "'headers' must be a schema",
+  severity: 1,
+  linterFunction: 'apilintElementOrClass',
+  linterParams: ['schema'],
+  marker: 'value',
+  target: 'headers',
+  data: {},
+};
+
+export default httpMessageBindingHeadersLint;
