+++ packages/apidom-ls/src/config/asyncapi/bindings/http/message-binding/lint/index.ts
@@ -1,0 +1,11 @@
+import httpMessageBindingAllowedFieldsLint from './allowed-fields';
+import httpMessageBindingHeadersLint from './headers';
+import httpMessageBindingsVersionLint from './bindingVersion';
+
+const lints = [
+  httpMessageBindingAllowedFieldsLint,
+  httpMessageBindingHeadersLint,
+  httpMessageBindingsVersionLint,
+];
+
+export default lints;
