+++ packages/apidom-ls/src/config/asyncapi/bindings/http/message-binding/meta.ts
@@ -1,7 +1,7 @@
 import lint from './lint';
 import completion from './completion';
 import documentation from './documentation';
-import { FormatMeta } from '../../../apidom-language-types';
+import { FormatMeta } from '../../../../../apidom-language-types';
 
 const meta: FormatMeta = {
   lint,
