+++ packages/apidom-ls/src/config/asyncapi/bindings/http/server-binding/lint/allowed-fields.ts
@@ -1,10 +1,10 @@
-import ApilintCodes from '../../../codes';
-import { LinterMeta } from '../../../../apidom-language-types';
+import ApilintCodes from '../../../../../codes';
+import { LinterMeta } from '../../../../../../apidom-language-types';
 
 const httpServerBindingAllowedFieldsLint: LinterMeta = {
   code: ApilintCodes.ALL_NOT_ALLOWED_FIELDS,
   source: 'apilint',
-  message: 'Object includes not allowed fields',
+  message: 'This object MUST NOT contain any properties. Its name is reserved for future use.',
   severity: 1,
   linterFunction: 'allowedFields',
   linterParams: [[]],
