+++ packages/apidom-ls/src/config/asyncapi/bindings/http/server-binding/meta.ts
@@ -1,5 +1,5 @@
 import lint from './lint';
-import { FormatMeta } from '../../../apidom-language-types';
+import { FormatMeta } from '../../../../../apidom-language-types';
 
 const meta: FormatMeta = {
   lint,
