+++ packages/apidom-ls/src/config/asyncapi/bindings/kafka/message-binding/documentation.ts
