+++ packages/apidom-ls/src/config/asyncapi/bindings/kafka/message-binding/lint/bindingVersion.ts
@@ -1,8 +1,8 @@
-import ApilintCodes from '../../../codes';
-import { LinterMeta } from '../../../../apidom-language-types';
+import ApilintCodes from '../../../../../codes';
+import { LinterMeta } from '../../../../../../apidom-language-types';
 
 const bindingVersionLint: LinterMeta = {
-  code: ApilintCodes.MESSAGE_BINDING_HTTP_BINDINGVERSION,
+  code: ApilintCodes.MESSAGE_BINDING_KAFKA_BINDINGVERSION,
   source: 'apilint',
   message: "bindingVersion' value must be a string",
   severity: 1,
