+++ packages/apidom-ls/src/config/asyncapi/bindings/kafka/message-binding/meta.ts
@@ -1,7 +1,7 @@
 import lint from './lint';
 import documentation from './documentation';
 import completion from './completion';
-import { FormatMeta } from '../../../apidom-language-types';
+import { FormatMeta } from '../../../../../apidom-language-types';
 
 const meta: FormatMeta = {
   lint,
